Refactor app logic / structure; add documentation The core logic of the application has been moved from the 'main.rs' file to 'lib.rs'. The 'main.rs' file now simply acts as an entrypoint for the application.
The change in structure is also necessary in order to allow for the
black-box testing approach we are going to use. Since integration tests
in Rust are considered to be external to our crate, we must use the
hybrid bin/lib approach.
  - Reference: [Integration Tests](https://doc.rust-lang.org/book/ch11-03-test-organization.html#integration-tests)

Doc comments were prepended source code files to explain their purpose
and additional documentation was added to some functions. Documentation
is very important and is something that, going forward, will be
maintained and, possibly, enforced.