Add integration test and refactor main/lib code
Integration testing has been set up using the designated 'tests' folder
as per Rust's conventions. The first test to be created is for the
'health' endpoint. In order to achieve the black box testing style that
we wanted, we needed to pull in some dev-dependencies (tests only):
  - `actix-rs`: provides access to the `actix_rt::test` macro
  - `reqwest`: HTTP client for sending requests to the server
  - `tokio`: underlying runtime, used to spawn/start the 'test' server

To decouple the tests and source code, some changes were made to
existing functions.
  - `run` returns an instance of `Server` instead of a future
  - `main` handles the error instance (if the server fails to start)
    using the `?` operator.