Refactored `run` function to accept a TcpListener
Since our tests involve spawning multiple instances of our application,
or an `actix::dev::Server` instance, to be exact, the `run` function
needed to be refactored to allow for binding on a random port.

Binding to port 0 is a potential security risk, so the server still
binds to a pre-determined port (8000 at the moment) in the application's
`main` function.

Making this change allows greater flexibility in that the behavior of
our binary does not change while our test suite can now reliably spawn
server instances on arbitrary ports (binding to port 0 causes the
underlying operating system to find an available, open port).

NOTE: The `spawn_server` test setup function signature needed to be
changed to return a `String` containing the spawned application address.