Add `subscribe` endpoint, update 'justfile' and CI
Added the `/subscribe` endpoint to the application with a "dummy"
implementation. The current implementation does not yet parse incoming
POST requests as it should.

The 'justfile' was updated to explicitly load environment variables so
that they may be used in `just` commands and/or scripts.

A typo in one of the Github Actions workflows was fixed. Minor changes
mostly involving naming were made to CI scripts.