Add script to start up a test database with Docker
Added a simple bash script to spin up an instance of PostgreSQL using
Docker. The script uses environment variables to populate the values
required and runs the `sqlx database create` command, which creates a
new database using the contents of the `DATABASE_URL` environment
variable.

This approach feels very hacky. It's simple and works, but it's not
optimal. A better solution would involve Docker Compose. For now, this
bash script meets our needs.