Add module for config handling and refactor code
The settings module enables configuration handling. The application is
now able to read a pre-defined __YAML__ file and populate configuration
variables from it. This feature is made possible thanks to the `config`
crate that was also recently added.

If the `load` function fails, then the server will not start. This
ensures that all required values are set before the application starts.

The "routes" module was renamed to "handlers" for two main reasons:
  - "handlers" is more consistent with the `actix-rs` ecosystem
  - "handlers" describes the behavior and purpose of the module and its
    corresponding sub-modules
    - See: [Request Handlers](https://actix.rs/docs/handlers/)