Database connection setup with url builder
The application is capable of utilizing the [settings](src/settings.rs)
module to build a valid connection string, or database URL. This is
required in order to work properly with SQLx.

The `build_url` function adds the ability to use the `DATABASE_URL`
environment variable value for the connection URI. If that fails, the
application falls back to using the local [config
file](config/default.yaml).

Small changes were made to some existing code for clarity. The logic was
not changed.

[X] __All tests are now passing__.