Create database connection pool and update tests
A shared database connection pool was implemented using the `PgPool`
type from `sqlx`. `actix` allows us to easily create a thread-safe
connection pool by adding it to the application state.

Many minor fixes were made and the codebase was refactored to provide
more consistent naming and, in some cases, behavior.

The Github Actions 'general.yml' workflow contained a duplicated step in
one of the jobs; this was removed.

The `config` crate was pulling in the wrong version of `serde`, causing
the derive macros to break. By disabling the default features of
`config` and explicitly setting the feature ('yaml') that we need, all
versions of `serde` were made compatible.

Integration tests were updated to reflect the changes made to the
application and the helper utilities that were added now allow all tests
to run successfully.