Clean up CI workflows and merge security jobs
The CI workflows found within the '.github/workflows' folder lack
consistency and some things needed to be made more explicit. The naming
of steps for certain workflows will likely be favored in the future as
failures in the CI workflow should be easier to find and/or debug.

The security workflows were merged back into a single job as there is no
difference at all between the two workflows. The same exact job is
triggered on `push` and on `schedule`.

These changes were minor but help push us in the right direction as we
are about to add telemtry (logging/tracing) to our application.