Enhance app telemetry and create separate module
The telemetry is now based entirely upon the `tracing` crate instead of
relying on `env_logger` and `log`. The `tracing` crate provides a much
more useful abstraction than traditional logging.

The logic was moved to a separate 'telemetry' module to reduce the size
of 'src/main.rs' and maintain separation of concerns.