Improve application telemetry and update tests
Added better telemetry support to the application using the `tracing`
crate and it's associated ecosystem. The `instrument` procedural
attribute was added to the "subscriptions" functions. This helps make
the code cleaner and easier to maintain.

The integration tests were updated to include logging (tracing)
capabilities as well. This can be turned on and off via an environment
variable: "TEST_LOG".

General code cleanup and refactoring was performed to improve the
readability and maintainability of the codebase. More work needs to be
done in this area but the primary focus is to ship an MVP.