Add Docker, offline SQLx support and fix CI
The application is being prepared for continuous deployment. A very
simple `Dockerfile` allows for the app to be containerized.

Problems with the CI pipeline revolving around `sqlx` should be fixed.
The '--locked' flag was removed as it was causing the CI workflows to
fail.

Minor changes were made to existing test names to make them consistent
and accurate.