Fix configuration and get CI to pass tests
CI jobs were failing due to a lack of a '.env' file in the root of the
repository. A workaround was attempted but it proved to be more trouble
than it is worth.

We will use a custom CD pipeline based on Docker Compose. This new and
improved setup should be included in one of the next commits.